Kasli: merge remoting into core_flash
