+++ artiq/core_flash.py
@@ -2,12 +2,15 @@ import os
 import sys
 import logging
 import tempfile
-import shutil
 import shlex
 import subprocess
 import hashlib
 import random
 import getpass
+import atexit
+
+from artiq import __artiq_dir__ as artiq_dir
+from artiq.frontend.bit2bin import bit2bin
 
 __all__ = ["LocalClient", "SSHClient"]
 
@@ -170,3 +173,75 @@ class SSHClient(Client):
 
     def run_command(self, cmd, **kws):
         self.drain(self.spawn_command(cmd, **kws))
+
+def build_dir (args):
+    bin_dir = args.dir
+    if bin_dir is None:
+        bin_dir = os.path.join(artiq_dir, "board-support")
+
+    needs_artifacts = not args.action or any(
+        action in args.action
+        for action in ["gateware", "rtm_gateware", "bootloader", "firmware", "load", "rtm_load"])
+    variant = args.variant
+    if needs_artifacts and variant is None:
+        variants = []
+        if args.srcbuild:
+            for entry in os.scandir(bin_dir):
+                if entry.is_dir():
+                    variants.append(entry.name)
+        else:
+            prefix = args.target + "-"
+            for entry in os.scandir(bin_dir):
+                if entry.is_dir() and entry.name.startswith(prefix):
+                    variants.append(entry.name[len(prefix):])
+        if args.target == "sayma":
+            try:
+                variants.remove("rtm")
+            except ValueError:
+                pass
+        if len(variants) == 0:
+            raise FileNotFoundError("no variants found, did you install a board binary package?")
+        elif len(variants) == 1:
+            variant = variants[0]
+        else:
+            raise ValueError("more than one variant found for selected board, specify -V. "
+                "Found variants: {}".format(" ".join(sorted(variants))))
+    variant_dir = None
+    rtm_variant_dir = None
+    if needs_artifacts:
+        if args.srcbuild:
+            variant_dir = variant
+        else:
+            variant_dir = args.target + "-" + variant
+        if args.target == "sayma":
+            if args.srcbuild:
+                rtm_variant_dir = "rtm"
+            else:
+                rtm_variant_dir = "sayma-rtm"
+    return (variant, bin_dir, variant_dir, rtm_variant_dir)
+
+def artifact_path(args, bin_dir, this_variant_dir, *path_filename):
+    if args.srcbuild:
+        # source tree - use path elements to locate file
+        return os.path.join(bin_dir, this_variant_dir, *path_filename)
+    else:
+        # flat tree - all files in the same directory, discard path elements
+        *_, filename = path_filename
+        return os.path.join(bin_dir, this_variant_dir, filename)
+
+def convert_gateware(bit_filename, header=False):
+    bin_handle, bin_filename = tempfile.mkstemp(
+        prefix="artiq_", suffix="_" + os.path.basename(bit_filename))
+    with open(bit_filename, "rb") as bit_file, \
+            open(bin_handle, "wb") as bin_file:
+        if header:
+            bin_file.write(b"\x00"*8)
+        bit2bin(bit_file, bin_file)
+        if header:
+            magic = 0x5352544d  # "SRTM", see sayma_rtm target
+            length = bin_file.tell() - 8
+            bin_file.seek(0)
+            bin_file.write(magic.to_bytes(4, byteorder="big"))
+            bin_file.write(length.to_bytes(4, byteorder="big"))
+    atexit.register(lambda: os.unlink(bin_filename))
+    return bin_filename
